Fill database with test data feature added
I use a list of nations and their respective capitals as test data. Clicking
the 'test' button will fill all these key-value pairs into the distributed
database. This means 236 entries. It is almost sure to overload one or more
instances, depending on the power of the machines hosting the pages, look out
for "timeout" and "left" messages.

The "put" functionality had to be factored out from the get/put button handler
function for this to work. The "get" function remains there, which will be
corrected later.