Uses random WSS/STUN servers from public lists
Bugout is all about eliminating any single point of failure, that is, trying to
achieve true serverless operation. Previously this was problematic since there
were only the 4 webtorrent trackers and 2 STUN servers which were hardwired
into Bugout as defaults. Now we use public server lists, sources are in
comments.

It would be nice to use DHT instead of trackers eventually, I opened
an issue about that on GitHub.

This version also adds a hacked instance of the minified Bugout library. This
is done to make it sure that no WSS tracker or STUN/TURN server defaults that
are hardwired into either Bugout or WebTorrent are used, only the ones supplied
by our code.